Scrub grade masters on first import
Graded sources should have the same groups and kerning as their ungraded
sibling source, as well as an identical font lib. 

Glyphs that are not present in the graded sources but in the ungraded
reference should be copied over as ungraded stand-ins. This currently
only works on first import of a new script; re-imports of the same
sources will see the copied glyphs and skip them.