Add /aaMatra-oriya to the "base" OpenType category
This glyph was specified as a base glyph in the vendor feature file, but
was not automatically placed in the OpenType category of the final UFOs.

From manual inspection in FontLab, the glyph does belong in the "base"
category, and so it has been manually added here.

If we encounter further issues where glyphs are missing their
appropriate category, we should investigate further.