Remove incorrect OpenType categories left over from first Sinhala import
In the original vendor sources, these glyphs were incorrectly classified
as marks. This was remedied, with subsequent vendor sources removing the
entries.

Despite this, our merge script did not remove the leftover entries in
the merged sources: it performs no cleanup for glyphs that the vendor
sources have no entry for.

These stale entries left shaping broken, even after the vendor had
provided a fix, and so this commit manually removes them. In the future,
we should adapt our merge script to avoid this issue.