Re-integrate improvements from Pixel repository
Furthers #650

This commit updates the scripts and subsets for Android subsetting with
the subsequent improvements that took place in the Pixel repository, as
part of moving the production of Android builds out of an external
repository and back in-repo.

Updated scripts and subsets taken from:

	googlefonts/pixel-brand-fallback-stack@73a9595cf7d477d647f28fbce44ca4df84f56545

Where necessary, modifications have been made to adapt the scripts back
to their new home.

The initial QA profile is to follow later, after #633 lands.

Co-authored-by: Ricky Atkins <ricky.atkins@daltonmaag.com>